+++ src/charm.py
@@ -6,6 +6,7 @@
 """Charm for Indico on kubernetes."""
 import logging
 import os
+import secrets
 import typing
 from re import findall
 from typing import Any, Dict, Iterator, List, Optional, Tuple
@@ -718,8 +719,15 @@ class IndicoOperatorCharm(CharmBase):  # pylint: disable=too-many-instance-attri
             container: Container where the plugins will be installed.
             plugins: List of plugins to be installed.
         """
-        if plugins:
-            install_command = ["pip", "install", "--upgrade"] + plugins
+        if not plugins:
+            return
+        # Fetch currently installed packages and push to a temp constraints file
+        constraints_file = f"/tmp/constraints-{secrets.token_hex(8)}.txt"  # nosec B108
+        current_packages, _ = container.exec(["pip", "freeze"]).wait_output()
+        container.push(constraints_file, current_packages, make_dirs=True)
+
+        try:
+            install_command = ["pip", "install", "--upgrade", "-c", constraints_file] + plugins
             logger.info("About to run: %s", " ".join(install_command))
             process = container.exec(
                 install_command,
@@ -727,6 +735,8 @@ class IndicoOperatorCharm(CharmBase):  # pylint: disable=too-many-instance-attri
             )
             output, _ = process.wait_output()
             logger.info("Output was: %s", output)
+        finally:  # Clean up constraints file
+            container.remove_path(constraints_file)
 
     def _get_indico_version(self) -> str:
         """Retrieve the current version of Indico.