+++ terraform/main.tf
@@ -2,7 +2,7 @@
 # See LICENSE file for licensing details.
 
 resource "juju_application" "indico" {
-  name  = var.app_name
+  name       = var.app_name
   model_uuid = var.model_uuid
 
   charm {