+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-01-05
+
+- Plugin installations now does not change installed Python packages.
+
 ## 2025-12-17
 
 - Moved charm-architecture.md from Explanation to Reference category.